refactoring on 'feature-coupled_stepDef' anti-pattern
